hardware: qcom: sm8150: display: fix -Wimplicit-int-float-conversion
All of the below have loss of precision due to implicit conversions to
float (or double). Accept the imprecision with explicit casts.

Fixes the following warnings from the compiler upgrade:

hardware/qcom/sm8150/display/sdm/libs/hwc2/hwc_display.cpp:760:45: error:
implicit conversion from 'uint32_t' (aka 'unsigned int') to 'float' may lose
precision [-Werror,-Wimplicit-int-float-conversion]
      layer->src_rect.right = layer_buffer->width;
                            ~ ~~~~~~~~~~~~~~^~~~~
hardware/qcom/sm8150/display/sdm/libs/hwc2/hwc_display.cpp:761:46: error:
implicit conversion from 'uint32_t' (aka 'unsigned int') to 'float' may lose
precision [-Werror,-Wimplicit-int-float-conversion]
      layer->src_rect.bottom = layer_buffer->height;
                             ~ ~~~~~~~~~~~~~~^~~~~~
hardware/qcom/sm8150/display/sdm/libs/core/display_base.cpp:271:56: error:
implicit conversion from 'uint32_t' (aka 'unsigned int') to 'float' may lose
precision [-Werror,-Wimplicit-int-float-conversion]
  if (gpu_target_layer_dst_xpixels > mixer_attributes_.width ||
                                   ~ ~~~~~~~~~~~~~~~~~~^~~~~
hardware/qcom/sm8150/display/sdm/libs/core/display_base.cpp:272:54: error:
implicit conversion from 'uint32_t' (aka 'unsigned int') to 'float' may lose
precision [-Werror,-Wimplicit-int-float-conversion]
    gpu_target_layer_dst_ypixels > mixer_attributes_.height) {
                                 ~ ~~~~~~~~~~~~~~~~~~^~~~~~
hardware/qcom/sm8150/display/sdm/libs/core/display_base.cpp:1309:42: error:
implicit conversion from 'uint32_t' (aka 'unsigned int') to 'float' may lose
precision [-Werror,-Wimplicit-int-float-conversion]
                                         layer_height), align_x);
                                         ^~~~~~~~~~~~
hardware/qcom/sm8150/display/sdm/libs/hwc2/hwc_display_pluggable_test.cpp:626:46:
error: implicit conversion from 'uint32_t' (aka 'unsigned int') to 'float' may
lose precision [-Werror,-Wimplicit-int-float-conversion]
  layer->src_rect = LayerRect(0, 0, var_info.x_pixels, var_info.y_pixels);
                    ~~~~~~~~~       ~~~~~~~~~^~~~~~~~
hardware/qcom/sm8150/display/sdm/libs/hwc2/hwc_display_pluggable_test.cpp:626:65:
error: implicit conversion from 'uint32_t' (aka 'unsigned int') to 'float' may
lose precision [-Werror,-Wimplicit-int-float-conversion]
  layer->src_rect = LayerRect(0, 0, var_info.x_pixels, var_info.y_pixels);
                    ~~~~~~~~~                          ~~~~~~~~~^~~~~~~~
hardware/qcom/sm8150/display/sdm/libs/hwc2/hwc_display_pluggable.cpp:178:37:
error: implicit conversion from 'uint32_t' (aka 'unsigned int') to 'float' may
lose precision [-Werror,-Wimplicit-int-float-conversion]
  uint32_t new_mixer_width = UINT32(mixer_width * FLOAT(1.0f - width_ratio));
                                    ^~~~~~~~~~~ ~
hardware/qcom/sm8150/display/sdm/include/utils/constants.h:43:43: note: expanded from macro 'UINT32'
                                          ^~~
hardware/qcom/sm8150/display/sdm/libs/hwc2/hwc_display_pluggable.cpp:179:38: error: implicit conversion from 'uint32_t' (aka 'unsigned int') to 'float' may lose precision [-Werror,-Wimplicit-int-float-conversion]
  uint32_t new_mixer_height = UINT32(mixer_height * FLOAT(1.0f - height_ratio));
                                     ^~~~~~~~~~~~ ~
hardware/qcom/sm8150/display/sdm/libs/core/drm/hw_device_drm.cpp:1764:57:
error: implicit conversion from 'uint32_t' (aka 'unsigned int') to 'float' may
lose precision [-Werror,-Wimplicit-int-float-conversion]
  float max_scale_up = hw_resource_.hw_dest_scalar_info.max_scale_up;
        ~~~~~~~~~~~~   ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~
hardware/qcom/sm8150/display/sdm/libs/core/fb/hw_device.cpp:1338:57: error:
implicit conversion from 'uint32_t' (aka 'unsigned int') to 'float' may lose
precision [-Werror,-Wimplicit-int-float-conversion]
  float max_scale_up = hw_resource_.hw_dest_scalar_info.max_scale_up;
        ~~~~~~~~~~~~   ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~
hardware/qcom/sm8150/display/sdm/libs/core/drm/hw_tv_drm.cpp:322:67: error:
implicit conversion from 'long' to 'float' may lose precision
[-Werror,-Wimplicit-int-float-conversion]
      float hdr_reset_time_start = ((hdr_reset_start.tv_sec*1000) + (hdr_reset_start.tv_usec/1000));
            ~~~~~~~~~~~~~~~~~~~~    ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
hardware/qcom/sm8150/display/sdm/libs/core/drm/hw_tv_drm.cpp:323:63: error:
implicit conversion from 'long' to 'float' may lose precision
[-Werror,-Wimplicit-int-float-conversion]
      float hdr_reset_time_end = ((hdr_reset_end.tv_sec*1000) + (hdr_reset_end.tv_usec/1000));
            ~~~~~~~~~~~~~~~~~~    ~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
hardware/qcom/sm8150/display/sdm/libs/core/strategy.cpp:176:47: error: implicit
conversion from 'uint32_t' (aka 'unsigned int') to 'float' may lose precision
[-Werror,-Wimplicit-int-float-conversion]
  float layer_mixer_width = mixer_attributes_.width;
        ~~~~~~~~~~~~~~~~~   ~~~~~~~~~~~~~~~~~~^~~~~
hardware/qcom/sm8150/display/sdm/libs/core/strategy.cpp:177:48: error: implicit
conversion from 'uint32_t' (aka 'unsigned int') to 'float' may lose precision
[-Werror,-Wimplicit-int-float-conversion]
  float layer_mixer_height = mixer_attributes_.height;
        ~~~~~~~~~~~~~~~~~~   ~~~~~~~~~~~~~~~~~~^~~~~~

Bug: 139945549
Test: mm

+++ sdm/libs/core/drm/hw_tv_drm.cpp
@@ -319,8 +319,8 @@ DisplayError HWTVDRM::UpdateHDRMetaData(HWLayers *hw_layers) {
     // metadata. This will be replaced with an idle timer implementation in the future.
     if (reset_hdr_flag) {
       gettimeofday(&hdr_reset_end, NULL);
-      float hdr_reset_time_start = ((hdr_reset_start.tv_sec*1000) + (hdr_reset_start.tv_usec/1000));
-      float hdr_reset_time_end = ((hdr_reset_end.tv_sec*1000) + (hdr_reset_end.tv_usec/1000));
+      float hdr_reset_time_start = FLOAT((hdr_reset_start.tv_sec*1000) + (hdr_reset_start.tv_usec/1000));
+      float hdr_reset_time_end = FLOAT((hdr_reset_end.tv_sec*1000) + (hdr_reset_end.tv_usec/1000));
 
       if (((hdr_reset_time_end-hdr_reset_time_start)/1000) >= MIN_HDR_RESET_WAITTIME) {
         memset(&hdr_metadata_, 0, sizeof(hdr_metadata_));
